Allow editor to support multiple files
