+++ app/css/components/file-editor.css
@@ -2,8 +2,4 @@
 
 .c-file-editor {
     height: 100%;
-
-    & > .c-file-editor-monaco {
-        height: 100%;
-    }
 }
